==> Data_Scientist_with_Python/02_Intermediate_Python/03-Logic_Control_Flow_Filtering.ipynb <==
## `Section 03: Logic, Control Flow and Filtering` 

### 01- Equality

* In the editor on the right, write code to see if `True` equals `False`.
* Write Python code to check if `-5 * 15` is not equal to `75`.
* Ask Python whether the strings `"pyscript" `and `"PyScript"` are equal.
* What happens if you compare booleans and integers? Write code to see if `True` and `1` are equal.

# Comparison of booleans
print(True == False)

# Comparison of integers
print(-5 * 15 != 75)

# Comparison of strings
print("pyscript" == "PyScript")

# Compare a boolean with an integer
print(True == 1)

### 02 Greater and less than
* Write Python expressions, wrapped in a `print()` function, to check whether:
    * `x` is greater than or equal to `-10`. `x` has already been defined for you.
    * `"test"` is less than or equal to `y`. `y` has already been defined for you.
    * `True` is greater than `False`.

# Comparison of integers
x = -3 * 6
print(x >= -10)

# Comparison of strings
y = "test"
print("test" <= y)

# Comparison of booleans
print(True > False)

### 03- Compare arrays
* Using comparison operators, generate boolean arrays that answer the following questions:
* Which areas in `my_house` are greater than or equal to `18`?
* You can also compare two NumPy arrays element-wise. Which areas in `my_house` are smaller than the ones in `your_house`?
* Make sure to wrap both commands in a `print()` statement so that you can inspect the output!

# Create arrays
import numpy as np
my_house = np.array([18.0, 20.0, 10.75, 9.50])
your_house = np.array([14.0, 24.0, 14.25, 9.0])

# my_house greater than or equal to 18
print(my_house >= 18)

# my_house less than your_house
print(my_house < your_house)


### 04- and, or, not (1)
* Write Python expressions, wrapped in a `print()` function, to check whether:
    * `my_kitchen` is bigger than 10 and smaller than 18.
    * `my_kitchen` is smaller than 14 or bigger than 17.
    * double the area of `my_kitchen` is smaller than triple the area of `your_kitchen`.

# import 
import numpy as np 

# Define variables
my_kitchen = 18.0
your_kitchen = 14.0

# my_kitchen bigger than 10 and smaller than 18?
print(my_kitchen > 10 and my_kitchen < 18)

# my_kitchen smaller than 14 or bigger than 17?
print(my_kitchen < 14 or my_kitchen > 17)


# Double my_kitchen smaller than triple your_kitchen?
print(2 * my_kitchen < 3 * your_kitchen)

### 05- and, or, not (2)

x = 8
y = 9
not(not(x < 3) and not(y > 14 or y > 10))

### 06- Boolean operators with NumPy

Before, the operational operators like `<` and `>=` worked with NumPy arrays out of the box. Unfortunately, this is not true for the boolean operators `and`, `or`, and `not`.
To use these operators with NumPy, you will need `np.logical_and()`, `np.logical_or()` and `np.logical_not()`. Here's an example on the `my_house` and `your_house` arrays from before to give you an idea:

* Generate boolean arrays that answer the following questions:
* Which areas in `my_house` are greater than `18.5` or smaller than `10`?
* Which areas are smaller than `11` in both `my_house` and `your_house`? Make sure to wrap both commands in `print()` statement, so that you can inspect the output.
 

# Create arrays
import numpy as np
my_house = np.array([18.0, 20.0, 10.75, 9.50])
your_house = np.array([14.0, 24.0, 14.25, 9.0])

# my_house greater than 18.5 or smaller than 10
print(np.logical_or(my_house > 18.5, my_house < 10))

# Both my_house and your_house smaller than 11
print(np.logical_and(my_house < 11, your_house<11))

### 07- if
* Examine the `if` statement that prints out `"looking around in the kitchen."` if `room` equals `"kit"`.
* Write another `if` statement that prints out "big place!" if `area` is greater than 15.


# Define variables
room = "kit"
area = 14.0

# if statement for room
if room == "kit" :
    print("looking around in the kitchen.")

# if statement for area
if area > 15: 
    print("big place!")

### 08- Add else
* Add an `else` statement to the second control structure so that "pretty small." is printed out if `area > 15` evaluates to `False`.




# Define variables
room = "kit"
area = 14.0

# if-else construct for room
if room == "kit" :
    print("looking around in the kitchen.")
else :
    print("looking around elsewhere.")

# if-else construct for area
if area > 15 :
    print("big place!")
else: 
    print("pretty small.")

### 09-Customize further: elif
* Add an `elif` to the second control structure such that "medium size, nice!" is printed out if `area` is greater than `10`.




# Define variables
room = "bed"
area = 14.0

# if-elif-else construct for room
if room == "kit" :
    print("looking around in the kitchen.")
elif room == "bed":
    print("looking around in the bedroom.")
else :
    print("looking around elsewhere.")

# if-elif-else construct for area
if area > 15 :
    print("big place!")

elif area > 10:
    print("medium size, nice!")

else :
    print("pretty small.")

### 10- Driving right (1)
* Extract the `drives_right` column as a Pandas Series and store it as `dr`.
* Use `dr`, a boolean Series, to subset the `cars` DataFrame. Store the resulting selection in `sel`.
* Print `sel`, and assert that `drives_right` is `True` for all observations.


# Import cars data
import pandas as pd
cars = pd.read_csv('data/cars.csv', index_col = 0)

# Extract drives_right column as Series: dr
dr = cars['drives_right']

# Use dr to subset cars: sel
sel = cars[dr]

# Print sel
print(sel)

### 11- Driving right (2)

Convert the code to a one-liner that calculates the variable `sel` as before.





# Import cars data
import pandas as pd
cars = pd.read_csv('data/cars.csv', index_col = 0)

# Convert code to a one-liner
sel = cars[cars['drives_right']]

# Print sel
print(sel)

### 12- Cars per capita (1)
* Select the `cars_per_cap` column from `cars` as a Pandas Series and store it as `cpc`.
* Use `cpc` in combination with a comparison operator and `500`. You want to end up with a boolean Series that's `True` if the corresponding country has a `cars_per_cap` of more than `500` and `False` otherwise. Store this boolean Series as `many_cars`.
* Use `many_cars` to subset `cars`, similar to what you did before. Store the result as `car_maniac`.
* Print out `car_maniac` to see if you got it right.


# Import cars data
import pandas as pd
cars = pd.read_csv('data/cars.csv', index_col = 0)

# Create car_maniac: observations that have a cars_per_cap over 500
cpc = cars['cars_per_cap']
many_cars = cpc > 500

# Print car_maniac
car_maniac = cars[many_cars]
print(car_maniac)

### 13- Cars per capita (2)
* Use the code sample provided to create a DataFrame `medium`, that includes all the observations of `cars` that have a `cars_per_cap` between `100` and `500`.
* Print out `medium`.

# Import cars data
import pandas as pd
cars = pd.read_csv('data/cars.csv', index_col = 0)

# Import numpy, you'll need this
import numpy as np

# Create medium: observations with cars_per_cap between 100 and 500
observations_range = np.logical_and(cars['cars_per_cap'] > 100, cars['cars_per_cap'] < 500)
medium = cars[observations_range]


# Print medium
print(medium)

==================================
### `The End`
==================================
